rename add_io_optical to add_fiber_array, add tapers to add_fiber_single
