+++ pp/components/waveguide.py
@@ -165,7 +165,7 @@ if __name__ == "__main__":
     # pp.show(c)
 
     # print(c.ports)
-    # cc = pp.routing.add_io_optical(c)
+    # cc = pp.routing.add_fiber_array(c)
     # pp.show(cc)
 
     # c = waveguide_slab()
