+++ pp/routing/__init__.py
@@ -8,7 +8,6 @@ from pp.routing.connect_bundle import (
     link_optical_ports,
 )
 from pp.routing.connect import connect_strip, connect_strip_way_points
-from pp.routing.connect_component import add_io_optical
 from pp.routing.manhattan import round_corners, route_manhattan
 from pp.routing.repackage import package_optical2x2
 from pp.routing.route_fiber_single import route_fiber_single
@@ -20,7 +19,6 @@ from pp.routing.add_electrical_pads_top import add_electrical_pads_top
 from pp.routing.add_electrical_pads_shortest import add_electrical_pads_shortest
 
 __all__ = [
-    "add_io_optical",
     "add_fiber_array",
     "connect_bundle",
     "connect_strip",
