+++ pp/routing/get_input_labels.py
@@ -1,9 +1,9 @@
-from pp.add_labels import get_input_label
+from typing import List
 from omegaconf.listconfig import ListConfig
 from phidl.device_layout import Label
+from pp.add_labels import get_input_label
 from pp.component import ComponentReference
 from pp.port import Port
-from typing import List
 
 
 def get_input_labels(
