+++ pp/samples/mask_custom/test_mask.py
@@ -24,7 +24,7 @@ def _route_filter(*args, **kwargs):
 
 
 def add_te(component, **kwargs):
-    c = pp.routing.add_io_optical(
+    c = pp.routing.add_fiber_array(
         component,
         grating_coupler=pp.c.grating_coupler_elliptical_te,
         route_filter=_route_filter,
@@ -36,7 +36,7 @@ def add_te(component, **kwargs):
 
 
 def add_tm(component, **kwargs):
-    c = pp.routing.add_io_optical(
+    c = pp.routing.add_fiber_array(
         component,
         grating_coupler=pp.c.grating_coupler_elliptical_tm,
         route_filter=_route_filter,
