+++ pp/test_properties/test_properties_containers_add_fiber_array_.yml
@@ -7,7 +7,7 @@ component:
   gap: 0.234
   with_elec_connections: true
 component_name: mzi2x2_WECTrue
-function_name: add_io_optical
+function_name: add_fiber_array
 gc_port_name: W0
 get_input_labels_function: get_input_labels
 get_route_factory: route_fiber_array
@@ -15,3 +15,4 @@ grating_coupler: grating_coupler_te
 kwargs: _empty
 name: mzi2x2_WECTrue_te
 taper_factory: taper
+taper_length: 10.0
